+++ CHANGELOG.md
@@ -101,6 +101,7 @@ Change categories:
   - `dec2bcd16()`: `pldm_bcd_dec2bcd16()`
   - `bcd2dec32()`: `pldm_bcd_bcd2dec32()`
   - `dec2bcd32()`: `pldm_bcd_dec2bcd32()`
+  - `ver2str()`: `pldm_base_ver2str()`
 
 ### Removed